Fix Vortice 3.8.3 API incompatibilities across Client rendering and audio
- SharpDXD3D11Manager: Add SharpGen.Runtime using; fix EnumAdapters1/EnumOutputs
  to use out-param form; fix GetFullscreenState out bool → RawBool; fix Present
  to cast to uint; rename CpuAccessFlags → CPUAccessFlags in Texture2D descriptors;
  replace 3-arg CreateRenderTargetView/CreateBitmapFromDxgiSurface with direct-return
  2-arg forms; cast RowPitch (uint) to int in all copy loops
- SharpDXD3D11SpriteRenderer: Cast ByteWidth to uint; rename CpuAccessFlags →
  CPUAccessFlags in BufferDescription; fix SamplerDescription field names
  (ComparisonFunc, MinLOD, MaxLOD); fix BlendState field names (BlendEnable,
  SourceBlendAlpha, DestinationBlendAlpha, BlendOperationAlpha, ColorWriteEnable cast)
- SharpDXD3D11RenderingPipeline: Replace RawMatrix5x4 object initializers with
  20-parameter constructors (CS0191 readonly field fix)
- SharpDXD3D9SpriteRenderer: Replace ReadOnlySpan<float> + SetVertexShaderConstant
  with direct Matrix4x4 overload; remove count arg from SetPixelShaderConstant calls
- SharpDXD3D9RenderingPipeline: Cast levelDesc.Width/Height to int for ?? operator
- DXSoundManager: Switch from Vortice.DirectSound.DirectSound.DirectSoundCreate8
  to DSound.DirectSoundCreate8() factory
- DXSound: Replace generic CreateSoundBuffer<T> with non-generic CreateSoundBuffer
  + QueryInterface; fix SetCurrentPosition/GetCurrentPosition int vs uint mismatch

+++ Client/Rendering/SharpDXD3D9/SharpDXD3D9RenderingPipeline.cs
@@ -341,8 +341,8 @@ namespace Client.Rendering.SharpDXD3D9
             finalTransform.M32 += translation.Y;
 
             var levelDesc = dxTexture.GetLevelDesc(0);
-            float width = sourceRectangle?.Width ?? levelDesc.Width;
-            float height = sourceRectangle?.Height ?? levelDesc.Height;
+            float width = sourceRectangle?.Width ?? (int)levelDesc.Width;
+            float height = sourceRectangle?.Height ?? (int)levelDesc.Height;
 
             if (TryDrawSpriteEffect(texture, new GdiRectangleF(0, 0, width, height), sourceRectangle, colour, finalTransform))
                 return;
